HTTP daemon support and several POCs
httpd.c provides a generic mechanisms for other plugins to register "url
listeners". It currently listens on port 6085 by default.

echo is a simple plugin to demonstrate/test httpd. It listens to /echo/ and
simply repeats back whatever the client said. That means that GET requests
will only warrant an empty reply.

status is similar to echo in it's PoC-nature. It listens to /status/ and
sends a varnishadm status-request and responds to the client with whatever
vadmin said.

vcl is slightly more clever. It currently listens to /vcl/ and handles GET
requests. For GET /vcl/ it will return 'vcl.list'. For GET /vcl/boot it
will respond with the boot vcl, for GET /vcl/foo, it will respond with the
foo-vcl etc. The idea is also to have POST or PUT to /vcl/ to upload VCL.