vadmin: Don't bother writing to a closed CLI
Not really a huge issue as such, but wrong nevertheless. The most obvious
example of this happening is if varnishd wasn't started and you tried using
the agent. cli_sock() would be called, but the connection would obviously
fail, but we'd try to write() anyway.

The assert() in cli_write() is a bit scary, but I don't see any way it
could trigger right now where it shouldn't.

Fixes #106