Update THANKS
Add a number of people from git log, remove descriptions (hard to maintain
over time), and add Sponsors-section.

Feel free to add your name or business here if you've made contributions
that are missed. Specially interested in Sponsors for those of you who are
listed as contributors, but did the work for a company (that is interested
in being listed).