ipc: Add some safety checks for multi-thread issues
The old vpush would've (correctly) triggered this.

Each IPC handle is thread-unique, and this test makes sure that this is
obeyed. If multi-thread support is needed in the future for HTTP, this test
will be valuable.

It's valuable today too, just a little bit less so.

The downside is the non-sense in http.c. I consider it worth it.