Add tests/multi.sh for thread-tests (not enabled)
This thing is a demonstration of how to abuse bash scripting.

The problem is simple: Parallel tests are needed if we ever want to do
multi-threading in http.c (Which would be neat, but there's a lot left to
do). This adds some idea of that.

I didn't enable it because it makes little sense, and it's ugly. But it's
there.