Retire parsing of Varnish 3 `param.show` output
It fixes a realloc-style leak of the `raw` parameter (`pos` in the call
site) by not duplicating the input for parsing. Consequently of the
Varnish 3 retirement the `get_name_val` function is now known as the
more accurate `parse_name`.

The `parse_name` function doesn't allocate the `param_opt` structure but
merely fills in the name field to be consistent with the `parse_value`
function. It also fixes another leak in `vparams_show_json` where the
very first `param_opt` allocation is always overwritten.

Refs #168